+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 
 * Improved handling of morphologically-analyzed words.
 * Include behaviors to restore to strings.
+* Include functions to write to internal bzip files.
+* (coverio)
 
 
 0.1.2 (2016-07-20)